primary logic for next number

+++ src/main/java/com/example/demo/dto/TaskDto.java
@@ -5,7 +5,7 @@ import lombok.Data;
 @Data
 public class TaskDto {
 	
-	Long oldValue;
-	Long newValue;
+	double oldValue;
+	double newValue;
 
 }
